% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticity_check.R
\name{authenticity_check}
\alias{authenticity_check}
\title{Check for authenticity of the analysis author or tutor from an indexed database}
\usage{
authenticity_check(name, affil)
}
\arguments{
\item{name}{Initials of the code's author name. (Last name, first name)}

\item{affil}{Affiliation of the code's author}
}
\value{
Text in console whether the statistical analysis is authorized by a tutor or not
}
\description{
Authenticity check for institutional analysis. Future implementation of this code is planned for incoming functions.
}
\examples{
\dontrun{ #For John Doe
authenticity_check(name = JD, affil = 'Some University')}
authenticity_check(name = 'FGJA', affil = 'UASLP')

}
