% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_bar_categ.R
\name{auto_bar_categ}
\alias{auto_bar_categ}
\title{auto_bar_categ}
\usage{
auto_bar_categ(data, groupvar, bar_args = list(), theme_func = theme_serene)
}
\arguments{
\item{data}{Name of the dataframe}

\item{groupvar}{Name of the grouping variable}

\item{bar_args}{List of arguments to be passed to "geom_bar"}

\item{theme_func}{Theme of the generated plots}
}
\description{
auto_bar_categ
}
