% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_bar_cont.R
\name{auto_bar_cont}
\alias{auto_bar_cont}
\title{Automatic barplot of continous variables}
\usage{
auto_bar_cont(
  data,
  groupvar,
  err_bar_show = TRUE,
  err_bar = c("sd", "se"),
  lang_labs = c("EN", "SPA")
)
}
\arguments{
\item{data}{Name of the dataframe}

\item{groupvar}{Grouping variable}

\item{err_bar_show}{Logical indicator. Default TRUE show error bars in columns. Default is TRUE}

\item{err_bar}{Statistic to be shown as error bar. Can be "sd" for standard deviation or "se" for standard error. Defauult is "se".}

\item{lang_labs}{Language of the resulting plots. Can be "EN" for english or "SPA" for spanish. Default is "SPA"}
}
\description{
Generates bar plots of contiuous variables based on numerical variables from a data frame.
}
