% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_pie_categ.R
\name{auto_pie_categ}
\alias{auto_pie_categ}
\title{Automatic generation of pieplots}
\usage{
auto_pie_categ(
  data,
  pie_bar_args = list(),
  theme_func = theme_serene_void,
  lang_labs = c("EN", "SPA"),
  statistics = TRUE,
  stat_lab = c("percent", "freq"),
  fill_grey = TRUE
)
}
\arguments{
\item{data}{Name of the dataframe}

\item{pie_bar_args}{List of arguments to be passed to "geom_bar"}

\item{theme_func}{Theme of the generated plots. Default is "theme_serene_void"}

\item{lang_labs}{Language of displayed labels. If null, default is spanish.}

\item{statistics}{Logical attribute to indicate if summary statistic parameters are shown.}

\item{stat_lab}{Statistics to be shown. Can choose if you want to show percentages or frequencies.}

\item{fill_grey}{Logical indicator to choose if the generated pie plots must be grey. Default is TRUE.}
}
\description{
Generates pie plots based on categorical variables of a data frame.
}
