% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_shapiro_raw.R
\name{auto_shapiro_raw}
\alias{auto_shapiro_raw}
\title{Automatic Shapiro-Wilk test table}
\usage{
auto_shapiro_raw(data)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}
}
\description{
Generates a HTML table of raw data from a numerical variables of a dataframe.
}
\examples{
auto_shapiro_raw(iris)
}
