% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_2g_2sid.R
\name{continuous_2g_2sid}
\alias{continuous_2g_2sid}
\title{Bivariate analysis for 2 groups}
\usage{
continuous_2g_2sid(data, groupvar)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}

\item{groupvar}{Grouping variable. Must have exactly 2 levels.}
}
\description{
Generates a HTML table of bivariate analysis for 2 groups.
}
\examples{
 # Not run
 # continuous_2g_2sid(dataframe, groupvar="group")
}
