% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_2g_pair.R
\name{continuous_2g_pair}
\alias{continuous_2g_pair}
\title{Bivariate analysis for 2 groups for paired data}
\usage{
continuous_2g_pair(
  data,
  groupvar,
  ttest_args = list(),
  wilcox_args = list(),
  flextableformat = TRUE
)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}

\item{groupvar}{Grouping variable. Must have exactly 2 levels.}

\item{ttest_args}{Arguments to be passed to \code{t.test()} function.}

\item{wilcox_args}{Arguments to be passed to \code{wilcox.test()} function.}

\item{flextableformat}{Logical operator to indicate the output desired. Default is TRUE. When FALSE, function will return a dataframe format.}
}
\value{
A dataframe or flextable with containing p values for paired tests along with statistics for normality and homocedasticity.
}
\description{
Generates a HTML table of bivariate paired analysis for 2 groups.
}
\examples{
data <- data.frame(group = rep(letters[1:2], 30),
                   var1 = rnorm(30, mean = 15, sd = 5),
                  var2 = rnorm(30, mean = 20, sd = 2),
                  var3 = rnorm(30, mean = 10, sd = 1),
                  var4 = rnorm(30, mean = 5, sd =2))
data$group<-as.factor(data$group)

continuous_2g_pair(data = data, groupvar = "group")


}
