% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_mixed_cols.R
\name{fix_mixed_cols}
\alias{fix_mixed_cols}
\title{Fix mixed cols in dataframes}
\usage{
fix_mixed_cols(df)
}
\arguments{
\item{df}{Dataframe to fix}
}
\value{
A dataframe with columns with columns without list
}
\description{
Fix mixed cols in dataframes
}
\examples{

col_1 <- c(1, 3, 4, 5, 5, 2)
col_2 <- c(1, 2, 3, 4, 5, 6)
col_3 <- c("a", "b", "c", "d", "e", "f")


df <- data.frame(col_1, col_2, col_3)

df$col_1 <- as.list(df$col_1) # Reproduce an importing error

str(df)

fixed_df <- fix_mixed_cols(df)
str(fixed_df)


}
