% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interp_continuous_2g.R
\name{interp_continuous_2g}
\alias{interp_continuous_2g}
\title{Generic interpretation for difference between 2 groups}
\usage{
interp_continuous_2g(data, groupvar, contvar)
}
\arguments{
\item{data}{Data frame}

\item{groupvar}{Grouping variable. Must have exact 2 levels}

\item{contvar}{Continuous variable to be contrasted}
}
\value{
It returns as output a generic text according to certain conditions which are as follows:
Format 1: Student's t-test with statistical significance. Conditions: Normality of model residuals, homogeneity of variances, p-value of the t-test<0.05.
Format 2: Student's t-test without statistical significance. Conditions: Normality of model residuals, homogeneity of variances, p-value>0.05.
Format 3: Welch's t-test with statistical significance. Conditions: Normality of model residuals, no homogeneity of variances, p-value<0.05.
Format 4: Welch's t-test without statistical significance. Conditions: Normality of model residuals, no homogeneity of variances, p-value>0.05.
Format 5: Mann-Withnney U test with statistical significance. Conditions: Non-normality of model residuals, p-value<0.05, independent of homogeneity of variances.
Format 6: Mann-Withnney U test without statistical significance. Conditions: Non-normality of model residuals, p-value>0.05, independent of homogeneity of variances.
}
\description{
Automatically writes a generic interpretation for the Student's t-test, Welch's t-test or Mann Whitney U-test. The wording only considers the alternative "two.sided" and the argument "paired" as "FALSE". Currently only the Spanish text is returned, more languages are planned to be added in the future.
}
