% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss_multreg.R
\name{ss_multreg}
\alias{ss_multreg}
\title{Sample Size Calculation for multiple regression analysis}
\usage{
ss_multreg(df, prop, logistic = FALSE)
}
\arguments{
\item{df}{Degrees of freedom planned to be introduced}

\item{prop}{Minimum prevalence of the expected event (Required if planned regression is a logistic regression)}

\item{logistic}{Logical operator to indicate wether the planned regression analysis is a logistic regression or not.}
}
\description{
Calculates the recommended sample size for a multiple regression analysis.
}
\examples{

ss_multreg(4, logistic = FALSE)
ss_multreg(4, prop = .6, logistic = TRUE)

}
