% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticity_check.R
\name{auth_check}
\alias{auth_check}
\alias{author_check}
\title{Author checker}
\usage{
author_check(name, affil_data, hash_col = "HASH")
}
\arguments{
\item{name}{Initials of the code's author name. (Last name, first name)}

\item{affil_data}{Dataframe with the hash code of the tutor.}

\item{hash_col}{Name of the column with hash codes. Default is "HASH"}
}
\description{
Author checker
}
