% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_dens_cont.R
\name{auto_dens_cont}
\alias{auto_dens_cont}
\title{auto_dens_cont}
\usage{
auto_dens_cont(
  data,
  s_mean = TRUE,
  s_median = TRUE,
  c_mean = "red",
  c_median = "blue",
  lt_mean = "solid",
  lt_median = "dotdash",
  lw_mean = 1,
  lw_median = 1,
  densplot_args = list(),
  theme_func = theme_serene
)
}
\arguments{
\item{data}{Name of the dataframe}

\item{s_mean}{Show mean. Logical operator to indicate if the mean should be plotted. Default is TRUE}

\item{s_median}{Show median. Logical operator to indicate if the median should be plotted. Default is TRUE}

\item{c_mean}{Color of the line indicating the mean. Default is red.}

\item{c_median}{Color of the line indicating the median. Default is blue.}

\item{lt_mean}{Linetype of the line indicating the mean. Default is "solid".}

\item{lt_median}{Linetype of the line indicating the median. Default is "dotdash".}

\item{lw_mean}{Linewidth of the line indicating the mean. Default is 1.}

\item{lw_median}{Linewidth of the line indicating the mean. Default is 1.}

\item{densplot_args}{List of arguments to be passed to "geom_density"}

\item{theme_func}{Theme to display plots. Default is "theme_serene"}
}
\value{
Returns a list containing the generated density plots
}
\description{
#' Automatically generates density plots of continuous variables from a database. The names of the variables are set to the names defined in the database. As a result, graphs generated with the default theme "theme_serene" will be obtained. In this function, the user must define each variable label with "label" function from "table1" package.
}
\author{
JMCR
}
