% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiVariAn-package.R
\docType{package}
\name{BiVariAn-package}
\alias{BiVariAn}
\alias{BiVariAn-package}
\title{BiVariAn: Bivariated Automatic Analysis}
\description{
Generates a automatic output for bivariate analysis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/AndresFloresG/BiVariAn}
  \item Report bugs at \url{https://github.com/AndresFloresG/BiVariAn/issues}
}

}
\keyword{internal}
