% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiVariAn_update.R
\name{BiVariAn_update}
\alias{BiVariAn_update}
\title{Update BiVariAn itself}
\usage{
BiVariAn_update(force = FALSE)
}
\arguments{
\item{force}{Whether to force an update, even if no newer version is available. (Currently even if FALSE, the package will update... working on process)}
}
\value{
Returns nothing else than an updated version of the package.
}
\description{
Use this function to update the development version of BiVariAn
}
\examples{
\dontrun{BiVariAn_update()}


}
