% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_2g_2sid.R
\name{continuous_2g_2sid}
\alias{continuous_2g_2sid}
\title{Bivariate analysis for 2 groups}
\usage{
continuous_2g_2sid(data, groupvar, flextableformat = TRUE)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}

\item{groupvar}{Grouping variable as character. Must have exactly 2 levels.}

\item{flextableformat}{Logical operator to indicate the output desired. Default is TRUE. When FALSE, function will return a dataframe format.}
}
\value{
Returns a dataframe or flextable of 2 groups 2 sided Mann Whitney's U or T test, along with Shapiro-Wilk's p values and Levene's p value.
}
\description{
Generates a HTML table of bivariate analysis for 2 groups.
}
\examples{
 data <- data.frame(group = rep(letters[1:2], 30),
 var1 = rnorm(30, mean = 15, sd = 5),
 var2 = rnorm(30, mean = 20, sd = 2),
 var3 = rnorm(30, mean = 10, sd = 1),
 var4 = rnorm(30, mean = 5, sd =2))

 data$group<-as.factor(data$group)

 conttable <- continuous_2g_2sid(data = data, groupvar = "group")
}
