% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_2g_pair.R
\name{continuous_2g_pair}
\alias{continuous_2g_pair}
\title{Bivariate analysis for 2 groups for paired data}
\usage{
continuous_2g_pair(data, groupvar, flextableformat)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}

\item{groupvar}{Grouping variable. Must have exactly 2 levels.}

\item{flextableformat}{Logical operator to indicate the output desired. Default is TRUE. When FALSE, function will return a dataframe format.}
}
\description{
Generates a HTML table of bivariate analysis for 2 groups.
}
