% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous_corr_test.R
\name{continuous_corr_test}
\alias{continuous_corr_test}
\title{Bivariate analysis for correlation tests}
\usage{
continuous_corr_test(
  data,
  referencevar,
  alternative = NULL,
  flextableformat = TRUE
)
}
\arguments{
\item{data}{Data frame from which variables will be extracted.}

\item{referencevar}{Reference variable. Must be a continuous variable.}

\item{alternative}{Alternative for cor.test. Must be either "two.sided", "geater" or "less"}

\item{flextableformat}{Logical operator to indicate the output desired. Default is TRUE. When FALSE, function will return a dataframe format.}
}
\description{
Automatic correlation analyses for continuous variables with one variable as reference.
}
