% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authenticity_check.R
\name{get_credentials}
\alias{get_credentials}
\title{Get credentials for author check}
\usage{
get_credentials(data = NULL, sep = ",")
}
\arguments{
\item{data}{Name of the file which the credentials are stored.}

\item{sep}{Separator of the dataframe. Default is "," to denote a .csv file.}
}
\value{
A dataframe with authorization credentials
}
\description{
Get credentials for author check
}
\examples{
\dontrun{
credentials<-get_credentials(data = "googlesheets_url", sep=",")}

}
