% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_serene.R
\name{theme_serene}
\alias{theme_serene}
\title{Basic theme for Bivaran packages plots}
\usage{
theme_serene(
  base_size = 14,
  base_family = "sans",
  base_fontface = "plain",
  base_line_size = base_size/14,
  base_rect_size = base_size/14,
  axis_text_angle = 0,
  border = F
)
}
\description{
Basic theme for Bivaran packages plots
}
\author{
Serene Chessur
}
