% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_serene.R
\name{theme_serene}
\alias{theme_serene}
\title{Basic theme for Bivaran packages plots}
\usage{
theme_serene(
  base_size = 14,
  base_family = "sans",
  base_fontface = "plain",
  base_line_size = base_size/14,
  base_rect_size = base_size/14,
  axis_text_angle = 0,
  border = F
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}

\item{base_fontface}{base font face}

\item{base_line_size}{base line size}

\item{base_rect_size}{base rect size}

\item{axis_text_angle}{Axis text angle}

\item{border}{Logical operator to indicate if the border should be printed}
}
\description{
Basic theme for Bivaran packages plots
}
\author{
Serene Chessur
}
